Move IntrinsicsOpt to separate file
