+++ src/opt.cpp
@@ -41,6 +41,7 @@
 #include "llvmutil.h"
 #include "module.h"
 #include "opt/ISPCPass.h"
+#include "opt/IntrinsicsOpt.h"
 #include "opt/ReplacePseudoMemoryOps.h"
 #include "sym.h"
 #include "util.h"
@@ -128,7 +129,6 @@ enum { BYTE = 1, WORD = 2, DWORD = 4, QWORD = 8, OWORD = 16, GRF = 32 };
 
 using namespace ispc;
 
-static llvm::Pass *CreateIntrinsicsOptPass();
 static llvm::Pass *CreateInstructionSimplifyPass();
 static llvm::Pass *CreatePeepholePass();
 
@@ -570,299 +570,6 @@ void ispc::Optimize(llvm::Module *module, int optLevel) {
 }
 
 ///////////////////////////////////////////////////////////////////////////
-// IntrinsicsOpt
-
-/** This is a relatively simple optimization pass that does a few small
-    optimizations that LLVM's x86 optimizer doesn't currently handle.
-    (Specifically, MOVMSK of a constant can be replaced with the
-    corresponding constant value, BLENDVPS and AVX masked load/store with
-    either an 'all on' or 'all off' masks can be replaced with simpler
-    operations.
-
-    @todo The better thing to do would be to submit a patch to LLVM to get
-    these; they're presumably pretty simple patterns to match.
-*/
-class IntrinsicsOpt : public llvm::FunctionPass {
-  public:
-    IntrinsicsOpt() : FunctionPass(ID){};
-
-    llvm::StringRef getPassName() const { return "Intrinsics Cleanup Optimization"; }
-
-    bool runOnBasicBlock(llvm::BasicBlock &BB);
-
-    bool runOnFunction(llvm::Function &F);
-
-    static char ID;
-
-  private:
-    struct MaskInstruction {
-        MaskInstruction(llvm::Function *f) { function = f; }
-        llvm::Function *function;
-    };
-    std::vector<MaskInstruction> maskInstructions;
-
-    /** Structure that records everything we need to know about a blend
-        instruction for this optimization pass.
-     */
-    struct BlendInstruction {
-        BlendInstruction(llvm::Function *f, uint64_t ao, int o0, int o1, int of)
-            : function(f), allOnMask(ao), op0(o0), op1(o1), opFactor(of) {}
-        /** Function pointer for the blend instruction */
-        llvm::Function *function;
-        /** Mask value for an "all on" mask for this instruction */
-        uint64_t allOnMask;
-        /** The operand number in the llvm CallInst corresponds to the
-            first operand to blend with. */
-        int op0;
-        /** The operand number in the CallInst corresponding to the second
-            operand to blend with. */
-        int op1;
-        /** The operand in the call inst where the blending factor is
-            found. */
-        int opFactor;
-    };
-    std::vector<BlendInstruction> blendInstructions;
-
-    bool matchesMaskInstruction(llvm::Function *function);
-    BlendInstruction *matchingBlendInstruction(llvm::Function *function);
-};
-
-char IntrinsicsOpt::ID = 0;
-
-/** Given an llvm::Value, return true if we can determine that it's an
-    undefined value.  This only makes a weak attempt at chasing this down,
-    only detecting flat-out undef values, and bitcasts of undef values.
-
-    @todo Is it worth working harder to find more of these?  It starts to
-    get tricky, since having an undef operand doesn't necessarily mean that
-    the result will be undefined.  (And for that matter, is there an LLVM
-    call that will do this for us?)
- */
-static bool lIsUndef(llvm::Value *value) {
-    if (llvm::isa<llvm::UndefValue>(value))
-        return true;
-
-    llvm::BitCastInst *bci = llvm::dyn_cast<llvm::BitCastInst>(value);
-    if (bci)
-        return lIsUndef(bci->getOperand(0));
-
-    return false;
-}
-
-bool IntrinsicsOpt::runOnBasicBlock(llvm::BasicBlock &bb) {
-    DEBUG_START_PASS("IntrinsicsOpt");
-
-    // We can't initialize mask/blend function vector during pass initialization,
-    // as they may be optimized out by the time the pass is invoked.
-
-    // All of the mask instructions we may encounter.  Note that even if
-    // compiling for AVX, we may still encounter the regular 4-wide SSE
-    // MOVMSK instruction.
-    if (llvm::Function *ssei8Movmsk =
-            m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_sse2_pmovmskb_128))) {
-        maskInstructions.push_back(ssei8Movmsk);
-    }
-    if (llvm::Function *sseFloatMovmsk =
-            m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_sse_movmsk_ps))) {
-        maskInstructions.push_back(sseFloatMovmsk);
-    }
-    if (llvm::Function *__movmsk = m->module->getFunction("__movmsk")) {
-        maskInstructions.push_back(__movmsk);
-    }
-    if (llvm::Function *avxFloatMovmsk =
-            m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_movmsk_ps_256))) {
-        maskInstructions.push_back(avxFloatMovmsk);
-    }
-
-    // And all of the blend instructions
-    blendInstructions.push_back(BlendInstruction(
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_sse41_blendvps)), 0xf, 0, 1, 2));
-    blendInstructions.push_back(BlendInstruction(
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_blendv_ps_256)), 0xff, 0, 1, 2));
-
-    llvm::Function *avxMaskedLoad32 =
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_maskload_ps_256));
-    llvm::Function *avxMaskedLoad64 =
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_maskload_pd_256));
-    llvm::Function *avxMaskedStore32 =
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_maskstore_ps_256));
-    llvm::Function *avxMaskedStore64 =
-        m->module->getFunction(llvm::Intrinsic::getName(llvm::Intrinsic::x86_avx_maskstore_pd_256));
-
-    bool modifiedAny = false;
-restart:
-    for (llvm::BasicBlock::iterator iter = bb.begin(), e = bb.end(); iter != e; ++iter) {
-        llvm::CallInst *callInst = llvm::dyn_cast<llvm::CallInst>(&*iter);
-        if (callInst == NULL || callInst->getCalledFunction() == NULL)
-            continue;
-
-        BlendInstruction *blend = matchingBlendInstruction(callInst->getCalledFunction());
-        if (blend != NULL) {
-            llvm::Value *v[2] = {callInst->getArgOperand(blend->op0), callInst->getArgOperand(blend->op1)};
-            llvm::Value *factor = callInst->getArgOperand(blend->opFactor);
-
-            // If the values are the same, then no need to blend..
-            if (v[0] == v[1]) {
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, v[0]);
-                modifiedAny = true;
-                goto restart;
-            }
-
-            // If one of the two is undefined, we're allowed to replace
-            // with the value of the other.  (In other words, the only
-            // valid case is that the blend factor ends up having a value
-            // that only selects from the defined one of the two operands,
-            // otherwise the result is undefined and any value is fine,
-            // ergo the defined one is an acceptable result.)
-            if (lIsUndef(v[0])) {
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, v[1]);
-                modifiedAny = true;
-                goto restart;
-            }
-            if (lIsUndef(v[1])) {
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, v[0]);
-                modifiedAny = true;
-                goto restart;
-            }
-
-            MaskStatus maskStatus = GetMaskStatusFromValue(factor);
-            llvm::Value *value = NULL;
-            if (maskStatus == MaskStatus::all_off) {
-                // Mask all off -> replace with the first blend value
-                value = v[0];
-            } else if (maskStatus == MaskStatus::all_on) {
-                // Mask all on -> replace with the second blend value
-                value = v[1];
-            }
-
-            if (value != NULL) {
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, value);
-                modifiedAny = true;
-                goto restart;
-            }
-        } else if (matchesMaskInstruction(callInst->getCalledFunction())) {
-            llvm::Value *factor = callInst->getArgOperand(0);
-            uint64_t mask;
-            if (GetMaskFromValue(factor, &mask) == true) {
-                // If the vector-valued mask has a known value, replace it
-                // with the corresponding integer mask from its elements
-                // high bits.
-                llvm::Value *value = (callInst->getType() == LLVMTypes::Int32Type) ? LLVMInt32(mask) : LLVMInt64(mask);
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, value);
-                modifiedAny = true;
-                goto restart;
-            }
-        } else if (callInst->getCalledFunction() == avxMaskedLoad32 ||
-                   callInst->getCalledFunction() == avxMaskedLoad64) {
-            llvm::Value *factor = callInst->getArgOperand(1);
-            MaskStatus maskStatus = GetMaskStatusFromValue(factor);
-            if (maskStatus == MaskStatus::all_off) {
-                // nothing being loaded, replace with undef value
-                llvm::Type *returnType = callInst->getType();
-                Assert(llvm::isa<llvm::VectorType>(returnType));
-                llvm::Value *undefValue = llvm::UndefValue::get(returnType);
-                llvm::ReplaceInstWithValue(iter->getParent()->getInstList(), iter, undefValue);
-                modifiedAny = true;
-                goto restart;
-            } else if (maskStatus == MaskStatus::all_on) {
-                // all lanes active; replace with a regular load
-                llvm::Type *returnType = callInst->getType();
-                Assert(llvm::isa<llvm::VectorType>(returnType));
-                // cast the i8 * to the appropriate type
-                llvm::Value *castPtr =
-                    new llvm::BitCastInst(callInst->getArgOperand(0), llvm::PointerType::get(returnType, 0),
-                                          llvm::Twine(callInst->getArgOperand(0)->getName()) + "_cast", callInst);
-                LLVMCopyMetadata(castPtr, callInst);
-                int align;
-                if (g->opt.forceAlignedMemory)
-                    align = g->target->getNativeVectorAlignment();
-                else
-                    align = callInst->getCalledFunction() == avxMaskedLoad32 ? 4 : 8;
-#if ISPC_LLVM_VERSION < ISPC_LLVM_11_0
-                llvm::Instruction *loadInst =
-                    new llvm::LoadInst(castPtr, llvm::Twine(callInst->getArgOperand(0)->getName()) + "_load",
-                                       false /* not volatile */, llvm::MaybeAlign(align), (llvm::Instruction *)NULL);
-#else
-                llvm::Instruction *loadInst = new llvm::LoadInst(
-                    returnType, castPtr, llvm::Twine(callInst->getArgOperand(0)->getName()) + "_load",
-                    false /* not volatile */, llvm::MaybeAlign(align).valueOrOne(), (llvm::Instruction *)NULL);
-#endif
-                LLVMCopyMetadata(loadInst, callInst);
-                llvm::ReplaceInstWithInst(callInst, loadInst);
-                modifiedAny = true;
-                goto restart;
-            }
-        } else if (callInst->getCalledFunction() == avxMaskedStore32 ||
-                   callInst->getCalledFunction() == avxMaskedStore64) {
-            // NOTE: mask is the 2nd parameter, not the 3rd one!!
-            llvm::Value *factor = callInst->getArgOperand(1);
-            MaskStatus maskStatus = GetMaskStatusFromValue(factor);
-            if (maskStatus == MaskStatus::all_off) {
-                // nothing actually being stored, just remove the inst
-                callInst->eraseFromParent();
-                modifiedAny = true;
-                goto restart;
-            } else if (maskStatus == MaskStatus::all_on) {
-                // all lanes storing, so replace with a regular store
-                llvm::Value *rvalue = callInst->getArgOperand(2);
-                llvm::Type *storeType = rvalue->getType();
-                llvm::Value *castPtr =
-                    new llvm::BitCastInst(callInst->getArgOperand(0), llvm::PointerType::get(storeType, 0),
-                                          llvm::Twine(callInst->getArgOperand(0)->getName()) + "_ptrcast", callInst);
-                LLVMCopyMetadata(castPtr, callInst);
-
-                int align;
-                if (g->opt.forceAlignedMemory)
-                    align = g->target->getNativeVectorAlignment();
-                else
-                    align = callInst->getCalledFunction() == avxMaskedStore32 ? 4 : 8;
-                llvm::StoreInst *storeInst = new llvm::StoreInst(rvalue, castPtr, (llvm::Instruction *)NULL,
-                                                                 llvm::MaybeAlign(align).valueOrOne());
-                LLVMCopyMetadata(storeInst, callInst);
-                llvm::ReplaceInstWithInst(callInst, storeInst);
-
-                modifiedAny = true;
-                goto restart;
-            }
-        }
-    }
-
-    DEBUG_END_PASS("IntrinsicsOpt");
-
-    return modifiedAny;
-}
-
-bool IntrinsicsOpt::runOnFunction(llvm::Function &F) {
-
-    llvm::TimeTraceScope FuncScope("IntrinsicsOpt::runOnFunction", F.getName());
-    bool modifiedAny = false;
-    for (llvm::BasicBlock &BB : F) {
-        modifiedAny |= runOnBasicBlock(BB);
-    }
-    return modifiedAny;
-}
-
-bool IntrinsicsOpt::matchesMaskInstruction(llvm::Function *function) {
-    for (unsigned int i = 0; i < maskInstructions.size(); ++i) {
-        if (maskInstructions[i].function != NULL && function == maskInstructions[i].function) {
-            return true;
-        }
-    }
-    return false;
-}
-
-IntrinsicsOpt::BlendInstruction *IntrinsicsOpt::matchingBlendInstruction(llvm::Function *function) {
-    for (unsigned int i = 0; i < blendInstructions.size(); ++i) {
-        if (blendInstructions[i].function != NULL && function == blendInstructions[i].function) {
-            return &blendInstructions[i];
-        }
-    }
-    return NULL;
-}
-
-static llvm::Pass *CreateIntrinsicsOptPass() { return new IntrinsicsOpt; }
-
-///////////////////////////////////////////////////////////////////////////
 
 /** This simple optimization pass looks for a vector select instruction
     with an all-on or all-off constant mask, simplifying it to the
@@ -1179,7 +886,7 @@ static llvm::Value *lGetBasePtrAndOffsets(llvm::Value *ptrs, llvm::Value **offse
 #endif
         llvm::Value *vec = shuffle->getOperand(1);
 
-        if (lIsUndef(vec) && llvm::isa<llvm::ConstantAggregateZero>(indices)) {
+        if (LLVMIsValueUndef(vec) && llvm::isa<llvm::ConstantAggregateZero>(indices)) {
             broadcastDetected = true;
         }
     }
