+++ src/opt/IntrinsicsOpt.h
@@ -1,0 +1,94 @@
+/*
+  Copyright (c) 2022, Intel Corporation
+  All rights reserved.
+
+  Redistribution and use in source and binary forms, with or without
+  modification, are permitted provided that the following conditions are
+  met:
+
+    * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+
+    * Redistributions in binary form must reproduce the above copyright
+      notice, this list of conditions and the following disclaimer in the
+      documentation and/or other materials provided with the distribution.
+
+    * Neither the name of Intel Corporation nor the names of its
+      contributors may be used to endorse or promote products derived from
+      this software without specific prior written permission.
+
+
+   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+   IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+   TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+   PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER
+   OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+   EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+   PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+   PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+   LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+   NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+
+#pragma once
+
+#include "ISPCPass.h"
+
+namespace ispc {
+/** This is a relatively simple optimization pass that does a few small
+    optimizations that LLVM's x86 optimizer doesn't currently handle.
+    (Specifically, MOVMSK of a constant can be replaced with the
+    corresponding constant value, BLENDVPS and AVX masked load/store with
+    either an 'all on' or 'all off' masks can be replaced with simpler
+    operations.
+
+    @todo The better thing to do would be to submit a patch to LLVM to get
+    these; they're presumably pretty simple patterns to match.
+*/
+class IntrinsicsOpt : public llvm::FunctionPass {
+  public:
+    IntrinsicsOpt() : FunctionPass(ID){};
+
+    llvm::StringRef getPassName() const { return "Intrinsics Cleanup Optimization"; }
+
+    bool runOnBasicBlock(llvm::BasicBlock &BB);
+
+    bool runOnFunction(llvm::Function &F);
+
+    static char ID;
+
+  private:
+    struct MaskInstruction {
+        MaskInstruction(llvm::Function *f) { function = f; }
+        llvm::Function *function;
+    };
+    std::vector<MaskInstruction> maskInstructions;
+
+    /** Structure that records everything we need to know about a blend
+        instruction for this optimization pass.
+     */
+    struct BlendInstruction {
+        BlendInstruction(llvm::Function *f, uint64_t ao, int o0, int o1, int of)
+            : function(f), allOnMask(ao), op0(o0), op1(o1), opFactor(of) {}
+        /** Function pointer for the blend instruction */
+        llvm::Function *function;
+        /** Mask value for an "all on" mask for this instruction */
+        uint64_t allOnMask;
+        /** The operand number in the llvm CallInst corresponds to the
+            first operand to blend with. */
+        int op0;
+        /** The operand number in the CallInst corresponding to the second
+            operand to blend with. */
+        int op1;
+        /** The operand in the call inst where the blending factor is
+            found. */
+        int opFactor;
+    };
+    std::vector<BlendInstruction> blendInstructions;
+
+    bool matchesMaskInstruction(llvm::Function *function);
+    BlendInstruction *matchingBlendInstruction(llvm::Function *function);
+};
+llvm::Pass *CreateIntrinsicsOptPass();
+} // namespace ispc
